four commit

+++ src/main/java/BlocgApplication/BlocgApplication/repositories/PostRepository.java
@@ -1,9 +1,10 @@
 package BlocgApplication.BlocgApplication.repositories;
 
+import BlocgApplication.BlocgApplication.Controllers.MainController;
 import BlocgApplication.BlocgApplication.models.entities.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PostRepository extends JpaRepository<Post, Integer> {
+public interface PostRepository extends JpaRepository<Post, Long> {
 }
